+++ DepRegAttributes.ExampleLibrary/RegisteredClasses/SingletonClassRegisteredAsSelf.cs
@@ -1,6 +1,6 @@
 ﻿namespace DepRegAttributes.ExampleLibrary.RegisteredClasses;
 
 [RegisterSingleton]
-public class SingletonClassRegigisteredAsSelf
+public class SingletonClassRegisteredAsSelf
 {
 }